redesigned the logic in async_process_channels_input() and forwarded_message(): Added channel fetching and user linking

+++ src/handlers/channels.py
@@ -200,7 +200,7 @@ async def process_show_channels_command(message: Message):
     user_id = message.from_user.id
     channels = await db.fetch_user_channels(user_id)
 
-    if channels is not None:
+    if channels:
         channel_names = [channel["channel_name"] for channel in channels]
         await message.answer(f"Ваши каналы:\n{', '.join(channel_names)}")
     else:
@@ -317,7 +317,7 @@ async def process_confirm_delete_callback(callback: CallbackQuery, state: FSMCon
 
     # Удаляем выбранные каналы из базы данных
     result = await db.delete_user_channels(user_id, selected_channels)
-    
+
     if result:
         await callback.message.edit_text(f"Каналы успешно удалены: {', '.join(selected_channels)}")
     else:
@@ -360,7 +360,7 @@ async def process_confirm_delete_all_callback(callback: CallbackQuery, state: FS
 
     # Удаляем все каналы из базы данных
     result = await db.clear_user_channels(user_id)
-    
+
     if result:
         await callback.message.edit_text("✅ Все каналы успешно удалены.")
     else:
@@ -467,8 +467,8 @@ async def stop_news_handler(message: Message, state: FSMContext):
 @router.message(F.forward_from_chat.func(lambda chat: chat and chat.type == 'channel'))
 async def handle_forwarded_message(message: Message, state: FSMContext):
     # Сбрасываем состояние, если есть активное
-    await state.clear()
-    
+    # await state.clear()
+
     if message.media_group_id:
         data = await state.get_data()
         processed_groups = data.get("processed_media_groups", set())
@@ -476,14 +476,13 @@ async def handle_forwarded_message(message: Message, state: FSMContext):
         if message.media_group_id in processed_groups:
             await message.delete()
             return
-        else:
-            processed_groups.add(message.media_group_id)
-            await state.update_data(processed_media_groups=processed_groups)
+
+        processed_groups.add(message.media_group_id)
+        await state.update_data(processed_media_groups=processed_groups)
     await forwarded_message(message)
 
-# Функция для обработки пересылки
 async def forwarded_message(message: Message):
-    
+
     user_id = message.from_user.id
     addition_timestamp = datetime.now().isoformat()
     channel = message.forward_from_chat.username
@@ -491,44 +490,35 @@ async def forwarded_message(message: Message):
 
     if not channel:
         await message.answer("❌ Канал не определен. Пожалуйста, убедитесь, что вы пересылаете сообщение из публичного канала.")
-        await message.delete()
         return
 
-    if not channel.startswith("@"):
-        channel = f"@{channel}"
+    channel = f"@{channel}" if not channel.startswith("@") else channel
 
     try:
-        tasks = [
-            asyncio.create_task(
-                summarizer.determine_channel_topic(
-                    await scraper.scrape_messages_long_term(channel, days=DAY_RANGE_INTERVAL, limit=10)
-                )
-            )
-        ]
+        exist_channel_id = await db.fetch_channel_id(channel)
+        if exist_channel_id:
+            linking_user_channel = await db.link_user_single_channel(user_id, exist_channel_id, addition_timestamp)
+            if linking_user_channel:
+                await message.answer(f"Канал {channel} успешно добавлен! ✔️\n\n Список ваших каналов - команда /show_channels")
+                await message.delete()
+            else:
+                await message.answer("Произошла ошибка при добавлении канала. Пожалуйста, попробуйте позже.")
 
-        # Ожидаем завершения всех задач
-        channel_topics = await asyncio.gather(*tasks)
-    except Exception as e:
-        # Присуждаем пустоту если выдает ошибку
-        channel_topics = []
-        logging.error("\nError determine_channel_topic for user %s: %s\n", user_id, e)
-
-    try:
-        success = await db.add_user_channels(user_id, [channel], addition_timestamp, channel_topics)
-        channels = await db.fetch_user_channels(user_id)
-        channels_names = ', '.join([channel["channel_name"] for channel in channels])
-
-        if success:
-            await message.answer(f"Канал {channel} успешно добавлен! ✔️ \nОбновленный список каналов: {channels_names}")
-            await message.delete()
         else:
-            await message.answer("Произошла ошибка при добавлении канала. Пожалуйста, попробуйте позже.")
-            await message.delete()
-            return
-    
+            messages = await scraper.scrape_messages_long_term(channel, days=DAY_RANGE_INTERVAL, limit=20)
+            channel_topic = await summarizer.determine_channel_topic(messages)
+
+            adding_channel = await db.add_single_channel(channel, channel_topic, addition_timestamp)
+            if adding_channel:
+                channel_id = await db.fetch_channel_id(channel)
+                await db.link_user_single_channel(user_id, channel_id, addition_timestamp)
+                await message.answer(f"Канал {channel} успешно добавлен! ✔️\n\n Список ваших каналов - команда /show_channels")
+                await message.delete()
+            else:
+                await message.answer("Ошибка при добавлении канала. Пожалуйста, попробуйте позже.")
     except Exception as e:
-        logging.error("\nError adding channels for user %s: %s\n", user_id, e)
-        await message.answer("Произошла ошибка при добавлении каналов. Попробуйте позже.")
+        logging.error("\nError adding channel for user %s: %s\n", user_id, e)
+        await message.answer("Произошла ошибка при добавлении канала. Пожалуйста, попробуйте позже.")
 
 ##################################### Обработка текста от юзера ####################################
 
@@ -565,35 +555,30 @@ async def async_process_channels_input(message: Message):
             "Пожалуйста, проверьте правильность написания и попробуйте снова."
         )
         return
-    
-    # Пробуем определить темы каналов
-    try:
-        tasks = [
-            asyncio.create_task(
-                summarizer.determine_channel_topic(
-                    await scraper.scrape_messages_long_term(channel, days=DAY_RANGE_INTERVAL, limit=10)
-                )
-            )
-            for channel in new_channels
-        ]
-
-        # Ожидаем завершения всех задач
-        channel_topics = await asyncio.gather(*tasks)
-    except Exception as e:
-        # Присуждаем пустоту если выдает ошибку
-        channel_topics = []
-        logging.error("\nError determine_channel_topic for user %s: %s\n", user_id, e)
 
     try:
-        channels = await db.fetch_user_channels(user_id)
-        channels_names = ', '.join([channel["channel_name"] for channel in channels])
+        new_channels = list(new_channels)
 
-        success = await db.add_user_channels(user_id, list(new_channels), addition_timestamp, channel_topics)
-        if success:
-            channels_list = ', '.join(new_channels)
-            await message.answer(f"Каналы успешно добавлены 👍\n{channels_list}. Обновленный список каналов: {channels_names}")
+        channel_ids = await db.fetch_channel_ids(new_channels)
+        if channel_ids:
+            await db.link_user_channels(user_id, channel_ids, addition_timestamp)
+            new_channels_list = ', '.join(new_channels)
+            await message.answer(f"Каналы {new_channels_list} успешно добавлены! ✔️\n\n Список ваших каналов - команда /show_channels")
+
         else:
-            await message.answer("Произошла ошибка при добавлении каналов. Попробуйте еще раз.")
+            topics = []
+            for channel in new_channels:
+                messages = await scraper.scrape_messages_long_term(channel, days=DAY_RANGE_INTERVAL, limit=20)
+                channel_topic = await summarizer.determine_channel_topic(messages)
+                topics.append(channel_topic)
+
+            await db.add_channels(new_channels, topics, addition_timestamp)
+            channel_ids = await db.fetch_channel_ids(new_channels)
+            await db.link_user_channels(user_id, channel_ids, addition_timestamp)
+
+            new_channels_list = ', '.join(new_channels)
+            await message.answer(f"Каналы {new_channels_list} успешно добавлены! ✔️\n\n Список ваших каналов - команда /show_channels")
+
     except Exception as e:
         logging.error("\nError adding channels for user %s: %s\n", user_id, e)
         await message.answer("Произошла ошибка при добавлении каналов. Попробуйте позже.")
